+++ ICD.Connect.Krang/Remote/Direct/API/AbstractApiMessage.cs
@@ -1,0 +1,10 @@
+﻿using ICD.Connect.API.Info;
+using ICD.Connect.Protocol.Network.Direct;
+
+namespace ICD.Connect.Krang.Remote.Direct.API
+{
+	public abstract class AbstractApiMessage : AbstractMessage
+	{
+		public ApiClassInfo Command { get; set; }
+	}
+}