+++ ICD.Connect.Krang/Remote/Direct/ShareDevices/ShareDevicesMessage.cs
@@ -6,7 +6,7 @@ using ICD.Connect.Routing.Endpoints.Destinations;
 using ICD.Connect.Routing.Endpoints.Groups;
 using ICD.Connect.Routing.Endpoints.Sources;
 
-namespace ICD.Connect.Krang.Remote.Direct
+namespace ICD.Connect.Krang.Remote.Direct.ShareDevices
 {
 	[Serializable]
 	public sealed class ShareDevicesMessage : AbstractMessage